feat: game page add business test

+++ game_page/game_page_data/src/main/kotlin/com/example/game_page_data/cache/models/GamePageEntity.kt
@@ -9,7 +9,7 @@ data class GamePageEntity(
     val id: String,
     val name: String?,
     val description: String?,
-    val rating: Long?,
+    val rating: Double?,
     val reditLink: String?,
     val reditDescription: String?,
     val platforms: List<PlatformEntity>?,
