+++ game_page/game_page_domain/src/main/kotlin/com/example/game_page_domain/models/GamePageModelUi.kt
@@ -4,7 +4,7 @@ data class GamePageModelUi(
     val id: String,
     val name: String?,
     val description: String?,
-    val rating: Long?,
+    val rating: Double?,
     val reditLink: String?,
     val reditDescription: String?,
     val platformUis: List<PlatformUi>?,
